Restructuring - Restructured the folders so that each example is in a separate folder, inside of which there are files and subfolders (eg. 'src' and 'images') related to the specific example. - Moved file 'util.ts' in the newly created folder 'common' since it is used in multiple examples. - Updated the path of 'util.ts' in the example source files. - Created a general README.md inside 'txwrapper-examples' folder that should give a short description about all examples and the folder structure. - Updated the examples paths in “tsconfig.build.json” and “package.json” files.
