+++ packages/txwrapper-examples/multisig/images/multisig-output-cli1.png
