+++ packages/txwrapper-examples/multisig/images/multisig-output-cli2.png
