+++ packages/txwrapper-examples/multisig/images/multisig-output-cli3.png
