+++ packages/txwrapper-examples/multisig/images/multisig-output-cli4.png
