+++ packages/txwrapper-examples/multisig/images/multisig-output-pjs-apps.png
