+++ packages/txwrapper-examples/package.json
@@ -14,10 +14,10 @@
   },
   "scripts": {
     "build": "yarn build:workspace",
-    "polkadot": "node lib/polkadot",
-    "polkadotBatchAll": "node lib/polkadotBatchAll",
+    "polkadot": "node lib/polkadot/src/polkadot",
+    "polkadotBatchAll": "node lib/polkadotBatchAll/src/polkadotBatchAll",
     "mandala": "node lib/mandala",
-    "multisig": "node lib/multisig"
+    "multisig": "node lib/multisig/src/multisig"
   },
   "dependencies": {
     "@polkadot/api": "8.3.2",
