+++ packages/txwrapper-examples/polkadotBatchAll/README.md
@@ -1,0 +1,4 @@
+# Polkadot Batch All example
+
+## Description
+Example that shows how to use the `batchAll` method.
